feat(admin): make /configs/validate available in all modes (#13220)
