+++ t/admin/standalone.spec.ts
@@ -574,7 +574,7 @@ describe('Admin - Standalone', () => {
       expect(resp.status).toEqual(400);
       expect(resp.data).toMatchObject({
         error_msg:
-          'invalid routes at index 0, err: invalid configuration: property "uri" validation failed: wrong type: expected string, got number',
+          'invalid configuration: property "uri" validation failed: wrong type: expected string, got number',
       });
     });
 
@@ -599,7 +599,7 @@ describe('Admin - Standalone', () => {
       expect(resp.status).toEqual(400);
       expect(resp.data).toEqual({
         error_msg:
-          'invalid services at index 0, err: unknown plugin [invalid-plugin]',
+          'unknown plugin [invalid-plugin]',
       });
       const resp2 = await clientException.put(
         ENDPOINT,
@@ -611,7 +611,7 @@ describe('Admin - Standalone', () => {
       expect(resp2.status).toEqual(400);
       expect(resp2.data).toEqual({
         error_msg:
-          'invalid routes at index 0, err: unknown plugin [invalid-plugin]',
+          'unknown plugin [invalid-plugin]',
       });
     });
 
@@ -626,7 +626,7 @@ describe('Admin - Standalone', () => {
       expect(resp.status).toEqual(400);
       expect(resp.data).toEqual({
         error_msg:
-          'invalid services at index 0, err: invalid configuration: failed to match pattern "^((uri|server_name|server_addr|request_uri|remote_port|remote_addr|query_string|host|hostname|mqtt_client_id)|arg_[0-9a-zA-z_-]+)$" with "args_invalid"',
+          'invalid configuration: failed to match pattern "^((uri|server_name|server_addr|request_uri|remote_port|remote_addr|query_string|host|hostname|mqtt_client_id)|arg_[0-9a-zA-z_-]+)$" with "args_invalid"',
       });
 
       const resp2 = await clientException.put(
@@ -639,7 +639,7 @@ describe('Admin - Standalone', () => {
       expect(resp2.status).toEqual(400);
       expect(resp2.data).toEqual({
         error_msg:
-          'invalid routes at index 0, err: invalid configuration: failed to match pattern "^((uri|server_name|server_addr|request_uri|remote_port|remote_addr|query_string|host|hostname|mqtt_client_id)|arg_[0-9a-zA-z_-]+)$" with "args_invalid"',
+          'invalid configuration: failed to match pattern "^((uri|server_name|server_addr|request_uri|remote_port|remote_addr|query_string|host|hostname|mqtt_client_id)|arg_[0-9a-zA-z_-]+)$" with "args_invalid"',
       });
     });
   });
@@ -784,7 +784,7 @@ describe('Validate API - Standalone', () => {
         errors: expect.arrayContaining([
           expect.objectContaining({
             resource_type: 'routes',
-            error: expect.stringContaining('invalid routes at index 0'),
+            error: expect.any(String),
           }),
         ]),
       });
